Align native surface continuation and portal path state with Cycles
Consume retained ShaderData/closures directly, preserve native labels and transparent-ray writes, and carry portal depth across surface/light/shadow contexts. Add original-HIP state and complete-render regressions, including corrected linked-position and shared-edge controls.

Validate host 178/178, HIP 190/190, fallback 192/192 and strict native Vulkan 9/9. Archive six 256-spp canaries and explicitly retain the unresolved Classroom finite gate, DiffInd residuals and Barbershop performance gap.

+++ tests/test_luisa_cycles_shadow_queue.cpp
@@ -140,7 +140,7 @@ bool run(const char *program, const char *backend, bool no_cache) {
             evaluator.trace_staged(task, params, active, visible);
             output.write(
                 id, make_float4(task.shadow_throughput, visible.cast<float>()));
-            meta.write(id, make_uint4(task.transparent_depth, task.rng_offset,
+            meta.write(id, make_uint4(task.transparent_depth.cast<unsigned>(), task.rng_offset,
                                       task.volume_bounds_bounce, id));
           }};
   auto output = device.create_buffer<luisa::float4>(paths);
